feat: Add support for ignored fields
Allow to show that certain input or output fields are known but are explicitly ignored.
This can be done using optional field (why would mandatory field be ignored?).
If only getter is present then it means input field is ignored.
If only setter is present then it means output field is ignored.
Those getters and setters would not be actually called.
They're just for documentation purposes.

This is useful when you want to explicitly document what you do with
each input or output field and want one-to-one mapping
between documentation and code.